fix: address PR review feedback (CodeRabbit + Claude)
Bugs fixed:
- Normalize model IDs in headless error path (model-validator.js)
- Wrap autoRepairAlias saveConfig in try/catch for best-effort persistence

File size violations resolved:
- bin/sidecar.js 434→283 lines (extracted start-helpers.js, cli-handlers.js)
- config.js 350→299 lines (extracted alias-resolver.js)
- api-key-store.test.js 659→3 files, all under 300 lines
- model-validator.test.js 344→2 files, all under 300 lines

Minor fixes:
- addAlias rejects 'null' name and trims whitespace
- Tightened config-null-alias.test.js assertions to check exact defaults
- Added un-mocked integration test for persisted key fallback
- Updated CLAUDE.md with normalizeModelId + 5 new test files
