+++ bin/sidecar.js
@@ -8,7 +8,6 @@
  */
 
 // Load environment variables from .env files
-const fs = require('fs');
 const path = require('path');
 require('dotenv').config({ path: path.join(__dirname, '..', '.env'), quiet: true });
 
@@ -22,7 +21,9 @@ const { syncOpenCodeAuth } = require('../src/utils/auth-sync');
 syncOpenCodeAuth();
 
 const { parseArgs, validateStartArgs, getUsage } = require('../src/cli');
-const { validateTaskId, safeSessionDir } = require('../src/utils/validators');
+const { validateTaskId } = require('../src/utils/validators');
+const { resolveModelFromArgs, validateFallbackModel } = require('../src/utils/start-helpers');
+const { handleSetup, handleAbort, handleUpdate, handleMcp } = require('../src/cli-handlers');
 
 const VERSION = require('../package.json').version;
 
@@ -111,52 +112,6 @@ async function main() {
     }
   } catch (err) {
     console.error(`Error: ${err.message}`);
-    process.exit(1);
-  }
-}
-
-/**
- * Resolve model from args: resolve alias or config default.
- * Returns { model, alias } or calls process.exit(1) on error.
- */
-function resolveModelFromArgs(args) {
-  const { resolveModel, loadConfig } = require('../src/utils/config');
-  const rawAlias = args.model;
-  let model;
-  try {
-    model = resolveModel(args.model);
-  } catch (err) {
-    console.error(err.message);
-    process.exit(1);
-  }
-
-  // Determine the alias used (explicit or config default)
-  let alias = rawAlias;
-  if (alias === undefined) {
-    const cfg = loadConfig();
-    if (cfg && cfg.default && !cfg.default.includes('/')) {
-      alias = cfg.default;
-    }
-  }
-  return { model, alias };
-}
-
-/**
- * Validate direct-API fallback models exist on the provider (opt-in via --validate-model).
- * Returns the (possibly corrected) model string.
- */
-async function validateFallbackModel(args, alias) {
-  const { detectFallback } = require('../src/utils/config');
-  if (!args['validate-model'] || !alias || !detectFallback(alias, args.model)) {
-    return args.model;
-  }
-  const { validateDirectModel } = require('../src/utils/model-validator');
-  try {
-    return await validateDirectModel(args.model, alias, {
-      headless: args['no-ui'] || !process.stdin.isTTY
-    });
-  } catch (err) {
-    console.error(err.message);
     process.exit(1);
   }
 }
@@ -319,112 +274,6 @@ async function handleRead(args) {
     metadata: args.metadata,
     project: args.cwd
   });
-}
-
-/**
- * Handle 'sidecar setup' command
- * Runs interactive setup wizard or adds an alias via --add-alias
- */
-async function handleSetup(args) {
-  const { addAlias, runInteractiveSetup, runApiKeySetup } = require('../src/sidecar/setup');
-
-  // Standalone API key window
-  if (args['api-keys']) {
-    const success = await runApiKeySetup();
-    if (success) {
-      console.log('API keys configured successfully.');
-    } else {
-      console.log('API key setup was not completed.');
-      process.exit(1);
-    }
-    return;
-  }
-
-  if (args['add-alias']) {
-    const spec = args['add-alias'];
-    const eqIndex = spec.indexOf('=');
-    if (eqIndex === -1) {
-      console.error('Error: --add-alias must be in format name=model');
-      process.exit(1);
-    }
-    const name = spec.slice(0, eqIndex);
-    const model = spec.slice(eqIndex + 1);
-    if (!name || !model) {
-      console.error('Error: --add-alias must be in format name=model');
-      process.exit(1);
-    }
-    addAlias(name, model);
-    console.log(`Alias '${name}' added: ${model}`);
-    return;
-  }
-
-  await runInteractiveSetup();
-}
-
-/**
- * Handle 'sidecar abort' command
- * Marks a running session as aborted
- */
-async function handleAbort(args) {
-  const taskId = args._[1];
-
-  if (!taskId) {
-    console.error('Error: task_id is required for abort');
-    console.error('Usage: sidecar abort <task_id>');
-    process.exit(1);
-  }
-
-  const taskIdCheck = validateTaskId(taskId);
-  if (!taskIdCheck.valid) {
-    console.error(taskIdCheck.error);
-    process.exit(1);
-  }
-
-  const project = args.cwd || process.cwd();
-  const sessionDir = safeSessionDir(project, taskId);
-  const metaPath = path.join(sessionDir, 'metadata.json');
-
-  if (!fs.existsSync(metaPath)) {
-    console.error(`Session ${taskId} not found`);
-    process.exit(1);
-  }
-
-  const meta = JSON.parse(fs.readFileSync(metaPath, 'utf-8'));
-  meta.status = 'aborted';
-  meta.abortedAt = new Date().toISOString();
-  fs.writeFileSync(metaPath, JSON.stringify(meta, null, 2), { mode: 0o600 });
-  console.log(`Session ${taskId} marked as aborted.`);
-}
-
-/**
- * Handle 'sidecar update' command
- * Updates claude-sidecar to the latest version
- */
-async function handleUpdate() {
-  const { performUpdate, getUpdateInfo, initUpdateCheck } = require('../src/utils/updater');
-  initUpdateCheck();
-  const info = getUpdateInfo();
-  if (info) {
-    console.log(`Updating claude-sidecar ${info.current} → ${info.latest}...`);
-  } else {
-    console.log('Updating claude-sidecar to latest...');
-  }
-  const result = await performUpdate();
-  if (result.success) {
-    console.log(`Updated successfully! Run 'sidecar --version' to verify.`);
-  } else {
-    console.error(`Update failed: ${result.error}`);
-    process.exit(1);
-  }
-}
-
-/**
- * Handle 'sidecar mcp' command
- * Starts the MCP server on stdio transport
- */
-async function handleMcp() {
-  const { startMcpServer } = require('../src/mcp-server');
-  await startMcpServer();
 }
 
 // Run main
