+++ tests/model-validator.test.js
@@ -240,64 +240,6 @@ describe('Model Validator', () => {
     });
   });
 
-  describe('normalizeModelId', () => {
-    let normalizeModelId;
-
-    beforeEach(() => {
-      const validator = require('../src/utils/model-validator');
-      normalizeModelId = validator.normalizeModelId;
-    });
-
-    it('should prepend provider when id lacks prefix', () => {
-      expect(normalizeModelId('google', 'gemini-3-flash')).toBe('google/gemini-3-flash');
-    });
-
-    it('should return as-is when id already has provider prefix', () => {
-      expect(normalizeModelId('google', 'google/gemini-3-flash')).toBe('google/gemini-3-flash');
-    });
-
-    it('should handle nested model ids (provider/org/model)', () => {
-      expect(normalizeModelId('openai', 'openai/gpt-4o')).toBe('openai/gpt-4o');
-    });
-
-    it('should prepend provider for bare model name', () => {
-      expect(normalizeModelId('anthropic', 'claude-sonnet-4.6')).toBe('anthropic/claude-sonnet-4.6');
-    });
-  });
-
-  describe('promptModelSelection saves normalized id', () => {
-    it('should normalize model id before saving to config', async () => {
-      const config = require('../src/utils/config');
-      const fetcher = require('../src/utils/model-fetcher');
-      const keyStore = require('../src/utils/api-key-store');
-
-      keyStore.readApiKeyValues.mockReturnValue({ google: 'test-key' });
-      // Models returned WITHOUT provider prefix (simulating provider API)
-      fetcher.fetchModelsFromProvider.mockResolvedValue([
-        { id: 'gemini-3-flash', name: 'Gemini 3 Flash' },
-        { id: 'gemini-3-pro', name: 'Gemini 3 Pro' },
-      ]);
-
-      const mockRl = { question: jest.fn(), close: jest.fn() };
-      mockRl.question.mockImplementation((_prompt, cb) => cb('1'));
-      jest.spyOn(require('readline'), 'createInterface').mockReturnValue(mockRl);
-      jest.spyOn(process.stderr, 'write').mockImplementation(() => true);
-
-      await validateDirectModel('google/gemini-old', 'gemini');
-
-      // Should save with provider prefix even though selected.id was bare
-      expect(config.saveConfig).toHaveBeenCalledWith(
-        expect.objectContaining({
-          aliases: expect.objectContaining({
-            gemini: 'google/gemini-3-flash',
-          }),
-        })
-      );
-
-      process.stderr.write.mockRestore();
-    });
-  });
-
   describe('filterRelevantModels', () => {
     it('should filter google models by gemini alias', () => {
       const result = filterRelevantModels(MOCK_GOOGLE_MODELS, 'gemini');
